Ajout du PemFactory dans VerifyPKCS7Signature

+++ class/VerifyPKCS7Signature.class.php
@@ -5,14 +5,24 @@ class VerifyPKCS7Signature {
     private $verifyPemCertificate;
     /** @var string */
     private $authorized_ca_path;
+    /** @var \PemCertificateFactory */
+    private $pemCertificateFactory;
 
 
-    public function __construct(string $authorized_ca_path, VerifyPemCertificateFactory $verifyPemCertificateFactory){
+    public function __construct(string $authorized_ca_path,
+                                VerifyPemCertificateFactory $verifyPemCertificateFactory,
+                                PemCertificateFactory $pemCertificateFactory
+    ){
         $this->verifyPemCertificate = $verifyPemCertificateFactory->get($authorized_ca_path);
         $this->authorized_ca_path = $authorized_ca_path;
+        $this->pemCertificateFactory = $pemCertificateFactory;
 	}
 
-	public function verify($file_path,$signature){
+    /**
+     * @throws \Exception
+     */
+    public function verify($file_path, $signature) : bool
+    {
 		try {
 			$signature_file = sys_get_temp_dir() . "/slow_signature_".mt_rand(0,mt_getrandmax());
 			$certificate_file = sys_get_temp_dir() . "/slow_certificate_".mt_rand(0,mt_getrandmax());
@@ -21,10 +31,10 @@ class VerifyPKCS7Signature {
 				
 		} catch(Exception $e){
 				
-			if (file_exists($certificate_file)) {
+			if (isset($certificate_file) && file_exists($certificate_file)) {
 				unlink($certificate_file);
 			}
-			if (file_exists($signature_file)){
+			if (isset($signature_file) && file_exists($signature_file)){
 				unlink($signature_file);
 			}
 			throw $e;
@@ -35,7 +45,10 @@ class VerifyPKCS7Signature {
 		return true;
 	}
 
-	private function verifyThrow($file_path,$signature,$signature_file,$certificate_file){
+    /**
+     * @throws \Exception
+     */
+    private function verifyThrow($file_path, $signature, $signature_file, $certificate_file){
 		$result = file_put_contents($signature_file, $signature);
 		if ($result === false){
 			throw new Exception("Impossible d'écrire la signature dans $signature_file");
@@ -48,6 +61,7 @@ class VerifyPKCS7Signature {
 		if ($result === false){
 			throw new Exception("Impossible d'écrire le certificat dans $certificate_file");
 		}
+        $this->pemCertificateFactory->getFromString($certificate)->checkCertificateIsValidAtDate(new DateTime());
 		$this->verifyPemCertificate->checkCertificateWithOpenSSL($certificate_file);
 		# On ne va pas vérifier le certificat (option -noverify)
         # Au niveau du purpose, smime est trop restrictif par rapport à notre besoin
@@ -61,13 +75,17 @@ class VerifyPKCS7Signature {
 		}
 	}
 
-	public function verifyCertificate($signature_content){
+    /**
+     * @throws \Exception
+     */
+    public function verifyCertificate(string $signature_content){
 		$signature_path = "/tmp/s2low_verify_pkcs7_".mt_rand(0,getrandmax());
 		file_put_contents($signature_path,$signature_content);
 		$certificate = $this->getCertificate($signature_path);
 		$certificate_path = "/tmp/s2low_verify_pkcs7_".mt_rand(0,getrandmax());
 		file_put_contents($certificate_path, $certificate);
 		try {
+		    $this->pemCertificateFactory->getFromString($certificate)->checkCertificateIsValidAtDate(new DateTime());
 			$this->verifyPemCertificate->checkCertificateWithOpenSSL(
 			    $certificate_path,
                 VerifyPemCertificate::CERTIFICATE_CHAIN_ERRORS
@@ -78,7 +96,8 @@ class VerifyPKCS7Signature {
 		}
 	}
 
-	private function getCertificate($signatureFileName){
+	private function getCertificate(string $signatureFileName) : string
+    {
 		$extractCmd = "openssl pkcs7 -in " . $signatureFileName . " -print_certs | openssl x509";
 
 		exec($extractCmd, $output, $ret);
